Update: With current code changes only `CastNullCheckDroppingsTest.java` test is failing.


contains s390x changes for ic-safepoint removal port. Build for slowdebug, fastdebug & release are working. But few of additional test failure are seen in tier-1 testing. 

```file
CastNullCheckDroppingsTest.java
SortingDeoptimizationTest.java
java/lang/StrictMath/ExhaustingTests.java 
java/math/BigInteger/LargeValueExceptions.java 
java/util/Random/RandomCanaryPi.java 
java/util/Spliterator/SpliteratorCollisions.java 
java/util/stream/test/org/openjdk/tests/java/util/SplittableRandomTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/CollectorsTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/CountLargeTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/FlatMapOpTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/InfiniteStreamWithLimitOpTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/MapOpTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/SliceOpTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/StreamLinkTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/StreamSpliteratorTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/ToArrayOpTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/WhileOpTest.java 
java/util/stream/test/org/openjdk/tests/java/util/stream/mapMultiOpTest.java 
java/util/zip/ZipFile/TestZipFileEncodings.java 
```



These changes needs to applied on top of https://github.com/openjdk/jdk/compare/master...fisk:jdk:remove-ic-safepoints-v2. 

Changes in diff file: 
[ic_safepoint_change.patch](https://github.com/openjdk/jdk/files/13668385/ic_safepoint_change.patch)

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [ ] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [ ] Commit message must refer to an issue

### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/17100/head:pull/17100` \
`$ git checkout pull/17100`

Update a local copy of the PR: \
`$ git checkout pull/17100` \
`$ git pull https://git.openjdk.org/jdk.git pull/17100/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 17100`

View PR using the GUI difftool: \
`$ git pr show -t 17100`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/17100.diff">https://git.openjdk.org/jdk/pull/17100.diff</a>

</details>