@TheRealMDoerr 2024-01-16
I think it's good. I recommend checking performance using some benchmark. Performance was very poor on PPC64 before this fix: https://github.com/fisk/jdk/commit/5aed08975dd9a9cb00acd4b4f4c071dfb962ac69
I didn't see any regression afterwards.