@TheRealMDoerr 2024-01-16
I have used SPEC benchmarks, but you can try any other one if you don't have a license. I don't remember which one you have used last time. I guess you will see a big difference if the fix is missing.