@offamitkumar 2024-01-16
> I have used SPEC benchmarks, but you can try any other one if you don't have a license. 

I guess we have license, So I'll try to access it.

> I guess you will see a big difference if the fix is missing.

Does that mean most of the benchmark will show the regression ? My understanding was that, different benchmarks are there to test different functionality.