@offamitkumar 2024-01-18
Hi Martin, 
I used renaissance benchmark and here is the result I got: 
with patch: 
```
Benchmark                                  Mode  Cnt      Score      Error  Units
o.r.actors.JmhAkkaUct.run                    ss   35   2705.773 ±  292.097  ms/op
o.r.actors.JmhReactors.run                   ss   15  20903.519 ± 1931.106  ms/op
o.r.jdk.concurrent.JmhFjKmeans.run           ss   40   4153.656 ±  386.497  ms/op
o.r.jdk.concurrent.JmhFutureGenetic.run      ss   65   3184.723 ±  125.662  ms/op
o.r.jdk.streams.JmhMnemonics.run             ss   25   8046.079 ±  314.166  ms/op
o.r.jdk.streams.JmhParMnemonics.run          ss   25   6528.264 ±  213.007  ms/op
o.r.jdk.streams.JmhScrabble.run              ss   65    277.482 ±   47.013  ms/op
o.r.neo4j.JmhNeo4jAnalytics.run              ss   30   2843.985 ±  103.831  ms/op
o.r.rx.JmhRxScrabble.run                     ss  105    217.778 ±   17.829  ms/op
o.r.scala.dotty.JmhDotty.run                 ss   65   1043.688 ±   67.138  ms/op
o.r.scala.sat.JmhScalaDoku.run               ss   30   7670.109 ±  756.502  ms/op
o.r.scala.stdlib.JmhScalaKmeans.run          ss   65    312.690 ±   15.646  ms/op
o.r.scala.stm.JmhPhilosophers.run            ss   40  10197.090 ±  716.933  ms/op
o.r.scala.stm.JmhScalaStmBench7.run          ss   80   4152.354 ±  123.389  ms/op
o.r.twitter.finagle.JmhFinagleChirper.run    ss  115  23160.800 ±  391.440  ms/op
o.r.twitter.finagle.JmhFinagleHttp.run       ss   20  10253.295 ± 1136.701  ms/op
```

without the patch: 
```
Benchmark                                  Mode  Cnt      Score      Error  Units
o.r.actors.JmhAkkaUct.run                    ss   35   2880.461 ±  296.531  ms/op
o.r.actors.JmhReactors.run                   ss   15  21792.433 ± 2346.775  ms/op
o.r.jdk.concurrent.JmhFjKmeans.run           ss   40   4259.036 ±  498.852  ms/op
o.r.jdk.concurrent.JmhFutureGenetic.run      ss   65   3199.238 ±  118.804  ms/op
o.r.jdk.streams.JmhMnemonics.run             ss   25   8488.351 ±  282.508  ms/op
o.r.jdk.streams.JmhParMnemonics.run          ss   25   7459.727 ±  202.225  ms/op
o.r.jdk.streams.JmhScrabble.run              ss   65    274.635 ±   45.646  ms/op
o.r.neo4j.JmhNeo4jAnalytics.run              ss   30   2959.414 ±   79.198  ms/op
o.r.rx.JmhRxScrabble.run                     ss  105    233.800 ±   19.422  ms/op
o.r.scala.dotty.JmhDotty.run                 ss   65   1053.025 ±   87.842  ms/op
o.r.scala.sat.JmhScalaDoku.run               ss   30   6547.708 ± 1425.924  ms/op
o.r.scala.stdlib.JmhScalaKmeans.run          ss   65    288.766 ±   16.716  ms/op
o.r.scala.stm.JmhPhilosophers.run            ss   40  10191.924 ±  564.472  ms/op
o.r.scala.stm.JmhScalaStmBench7.run          ss   80   4414.132 ±  163.167  ms/op
o.r.twitter.finagle.JmhFinagleChirper.run    ss  115  23875.359 ±  491.453  ms/op
o.r.twitter.finagle.JmhFinagleHttp.run       ss   20  12543.643 ±  600.994  ms/op
```


This is single-run result, and machine is shared-resource. I was only able to run it once (with & without patch) as it takes more than 6 hours to finish.